20200811 14:33 Repository Setting

+++ src/main/java/com/muffin/web/asset/AssetRepository.java
@@ -1,5 +1,5 @@
 package com.muffin.web.asset;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AssetRepository extends JpaRepository<Asset, Integer>, IAssetRepository{
+public interface AssetRepository extends JpaRepository<Asset, Long>, IAssetRepository{
 }
